Create Entry class, and rewrite tests
The entry class stores details of a withdrawal or deposit, and has
methods to return them formatted in the correct way for the statement.

Update Account and Printer classes to use the Entry class.

We now have a passing feature test for printing a statement on an
account with one deposit.