Remove transaction class and change Printer to Formatter
Big refactor! Should have been two commits really...

Remove Transaction class
========================

There were two problems with the Transaction class:
  1. It was formatting its data to go in the statement, e.g. money
  amounts formatted to 2dp. That presentation logic should be in a
  Printer/Formatter class.
  2. If you remove that logic from the Transaction class, then it's not
  doing anything - it's just storing state.

So I refactored the presentation logic (formatting amounts to 2dp and
formatting the date) into the Printer class, and then removed the
Transaction class altogether, replacing it with a Struct inside the
Account class. This is more appropriate than a Class for simply storing
state, and is better than using a hash as it prevents future developers
from adding in extra elements which we haven't defined.

Change Printer to Formatter
===========================

Previously I had a Printer module, with a method
Printer.print(transactions) which printed out the statement. This was
being called from Account.print_statement, which meant that that method
wasn't actually doing anything itself, which was making it very hard to
write a unit test for that method.

I've moved the responsibility for printing the statement into
Account.print_statement, and changed the Printer module into a Formatter
module, which returns the formatted statement as a string. This
separation has allowed me to write more meaningful unit tests for my
methods.